Move createDateCriteria
